chore(elements): Move tests to __tests__ folder
